requirements.txt
attrs
enum34
six
